Fix bugs in Qlearning training

+++ rllib/value_function/nn_value_function.py
@@ -64,16 +64,16 @@ class NNQFunction(AbstractQFunction):
                 return self._q_function(state)
             else:
                 in_ = torch.scatter(torch.zeros(self.num_states), 0, state, 1)
-                return self._q_function(in_).gather(1, action.unsqueeze(-1)).squeeze(-1)
+                return self._q_function(in_).gather(1, action.long().unsqueeze(-1)).squeeze(-1)
 
         if not self.discrete_state and not self.discrete_action:
             state_action = torch.cat((state, action), dim=-1)
             return self._q_function(state_action)
         elif self.discrete_state and self.discrete_action:
             in_ = torch.scatter(torch.zeros(self.num_states), 0, state, 1)
-            return self._q_function(in_).gather(1, action.unsqueeze(-1)).squeeze(-1)
+            return self._q_function(in_).gather(1, action.long().unsqueeze(-1)).squeeze(-1)
         elif not self.discrete_state and self.discrete_action:
-            return self._q_function(state).gather(1, action.unsqueeze(-1)).squeeze(-1)
+            return self._q_function(state).gather(1, action.long().unsqueeze(-1)).squeeze(-1)
         else:
             raise ValueError("If states are discrete, so should be actions.")
